Handle datatype error

+++ autotrading/data_engine/collector.py
@@ -89,7 +89,7 @@ class Collector:
                 'change_rate' : '전일종가대비변화량'
             }
             df = df.drop(columns=['timestamp', 'prev_closing_price']).rename(columns=columns)
-            df = df.astype({'누적거래금액':np.int128, '누적거래량':np.int128})
+            df = df.astype({'누적거래금액':np.int64, '누적거래량':np.int64})
             dtypes = self.check_dtypes(df)
             df = df.astype(dtypes)
             
@@ -106,7 +106,7 @@ class Collector:
                 'candle_acc_trade_volume' : '누적거래량'
             }
             df = df.drop(columns=['timestamp', 'unit']).rename(columns=columns)
-            df = df.astype({'누적거래금액':np.int128, '누적거래량':np.int128})
+            df = df.astype({'누적거래금액':np.int64, '누적거래량':np.int64})
             dtypes = self.check_dtypes(df)
             df = df.astype(dtypes)
